requirements.txt
django>=4.2.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.20.0
django-cors-headers>=4.0.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
scikit-learn>=1.3.0
numpy>=1.24.0 